Confirm delete

+++ src/app/usersmanagement/usersmanagement.component.ts
@@ -3,6 +3,7 @@ import { UserService } from '../services/user.service';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { UpdateComponent } from '../update/update.component';
+import { ConfirmDeleteComponent } from '../confirm-delete/confirm-delete.component'; // Importer le composant de confirmation
 import { User } from '../user.model';
 
 @Component({
@@ -11,15 +12,15 @@ import { User } from '../user.model';
   styleUrls: ['./usersmanagement.component.css'],
 })
 export class UsersmanagementComponent implements OnInit {
-  dataSource: MatTableDataSource<User> = new MatTableDataSource<User>(); // Initialise dataSource
+  dataSource: MatTableDataSource<User> = new MatTableDataSource<User>();
   displayedColumns: string[] = ['name', 'actions', 'status', 'update'];
-  users: User[] = []; // Liste des utilisateurs pour maintenir les données
-  filterValue: string = ''; // Valeur du filtre
+  users: User[] = [];
+  filterValue: string = '';
 
   constructor(private dialog: MatDialog, private userService: UserService) {}
 
   ngOnInit(): void {
-    this.tableData(); // Récupère les données au chargement du composant
+    this.tableData();
   }
 
   handleUpdateAction(user: User): void {
@@ -31,20 +32,19 @@ export class UsersmanagementComponent implements OnInit {
     const dialogConfig = new MatDialogConfig();
     dialogConfig.width = '550px';
     dialogConfig.disableClose = true;
-    dialogConfig.data = { user }; // Envoi de l'utilisateur actuel au modal
+    dialogConfig.data = { user };
 
     const dialogRef = this.dialog.open(UpdateComponent, dialogConfig);
 
     dialogRef.afterClosed().subscribe((updatedData: User) => {
       if (updatedData) {
-        // Assertion de type
         this.userService
           .updateUser(user.id as number, updatedData)
           .subscribe((updatedUser) => {
             const index = this.users.findIndex((u) => u.id === updatedUser.id);
             if (index !== -1) {
-              this.users[index] = updatedUser; // Met à jour la liste locale
-              this.dataSource.data = [...this.users]; // Rafraîchit dataSource
+              this.users[index] = updatedUser;
+              this.dataSource.data = [...this.users];
             }
           });
       }
@@ -60,7 +60,7 @@ export class UsersmanagementComponent implements OnInit {
     const newStatus = user.status === 'true' ? 'false' : 'true';
     this.userService.updateUserStatus(user.id, newStatus).subscribe({
       next: () => {
-        user.status = newStatus; // Met à jour le statut localement après succès
+        user.status = newStatus;
         this.dataSource.data = [...this.users];
       },
       error: (err) => {
@@ -70,27 +70,41 @@ export class UsersmanagementComponent implements OnInit {
   }
 
   deleteUser(id: number): void {
-    this.userService.deleteUserById(id).subscribe(
-      () => {
-        this.users = this.users.filter((user) => user.id !== id); // Supprime l'utilisateur localement
-        this.dataSource.data = [...this.users]; // Met à jour dataSource
-        console.log('Utilisateur supprimé avec succès.');
-      },
-      (error) => {
-        console.error("Erreur lors de la suppression de l'utilisateur", error);
+    // Ouvrir la boîte de dialogue de confirmation
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent, {
+      width: '350px',
+      disableClose: true,
+    });
+
+    // Traitement après la fermeture de la boîte de dialogue
+    dialogRef.afterClosed().subscribe((confirm: boolean) => {
+      if (confirm) {
+        this.userService.deleteUserById(id).subscribe(
+          () => {
+            this.users = this.users.filter((user) => user.id !== id);
+            this.dataSource.data = [...this.users];
+            console.log('Utilisateur supprimé avec succès.');
+          },
+          (error) => {
+            console.error(
+              "Erreur lors de la suppression de l'utilisateur",
+              error
+            );
+          }
+        );
       }
-    );
+    });
   }
 
   tableData(): void {
     this.userService.getAllUsers().subscribe((response: User[]) => {
       this.users = response;
-      this.dataSource.data = this.users; // Initialise dataSource avec les données
+      this.dataSource.data = this.users;
     });
   }
 
   applyFilter(event: Event): void {
     const filterValue = (event.target as HTMLInputElement).value;
-    this.dataSource.filter = filterValue.trim().toLowerCase(); // Met à jour le filtre de la table
+    this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 }
